refactor: 싱크 받은 후 커밋

+++ src/main/java/com/leafpage/controller/rental/RentalController.java
@@ -24,7 +24,7 @@ public class RentalController implements Controller {
         RentalDAO rentalDAO = RentalDAO.getInstance();
 
         String isbn = request.getParameter("ISBN");
-        Long userNo = ((Integer) session.getAttribute("userNo")).longValue();
+        Long userNo = (Long) session.getAttribute("userNo");
         log.debug("userNo = {}", userNo);
         log.debug("ISBN = {}", isbn);
